Reorganizing packages to cleanup
