+++ src/main/java/com/coveros/selenified/selenium/Is.java
@@ -5,8 +5,8 @@ import org.openqa.selenium.NoAlertPresentException;
 import org.openqa.selenium.WebDriver;
 import org.testng.log4testng.Logger;
 
-import com.coveros.selenified.output.OutputFile;
 import com.coveros.selenified.tools.General;
+import com.coveros.selenified.tools.OutputFile;
 
 public class Is {
