+++ src/main/java/com/coveros/selenified/services/Call.java
@@ -2,9 +2,9 @@ package com.coveros.selenified.services;
 
 import org.testng.log4testng.Logger;
 
-import com.coveros.selenified.output.OutputFile;
-import com.coveros.selenified.output.OutputFile.Result;
 import com.coveros.selenified.tools.General;
+import com.coveros.selenified.tools.OutputFile;
+import com.coveros.selenified.tools.OutputFile.Result;
 
 public class Call {
     private static final Logger log = Logger.getLogger(General.class);
