+++ src/main/java/com/coveros/selenified/services/Response.java
@@ -22,8 +22,8 @@ package com.coveros.selenified.services;
 
 import java.util.Map;
 
-import com.coveros.selenified.output.OutputFile;
-import com.coveros.selenified.output.OutputFile.Success;
+import com.coveros.selenified.tools.OutputFile;
+import com.coveros.selenified.tools.OutputFile.Success;
 import com.google.gson.JsonArray;
 import com.google.gson.JsonElement;
 import com.google.gson.JsonObject;
