+++ src/main/java/com/coveros/selenified/tools/Listener.java
@@ -24,8 +24,8 @@ import org.testng.ITestResult;
 import org.testng.Reporter;
 import org.testng.TestListenerAdapter;
 
-import com.coveros.selenified.output.OutputFile.Result;
 import com.coveros.selenified.selenium.Selenium.Browser;
+import com.coveros.selenified.tools.OutputFile.Result;
 
 import java.io.File;
